try fix types in public methods

+++ sql/orioledb--1.5--1.6_dev.sql
@@ -5,12 +5,12 @@
 
 -- Get current logical xid to remember it
 CREATE FUNCTION orioledb_get_current_logical_xid()
-RETURNS bigint
+RETURNS int8
 AS 'MODULE_PATHNAME'
 VOLATILE LANGUAGE C;
 
 -- Get current heap xid to remember it
 CREATE FUNCTION orioledb_get_current_heap_xid()
-RETURNS bigint
+RETURNS int8
 AS 'MODULE_PATHNAME'
 VOLATILE LANGUAGE C;
